Add files via upload
**Files Provided:**

1. **`companies.csv`**: A dataset containing details of S&P 500 companies.
    - **Columns**:
        - `Name`: Name of the company.
        - `Ticker`: Stock ticker symbol of the company.
        - `Sector`: Industry sector to which the company belongs.

2. **`stock_prices.csv`**: A dataset with daily stock prices for the past year for the companies listed in `companies.csv`.
    - **Columns**:
        - Company tickers derived from the `companies.csv` file.
    - **Note**: The values in this dataset may not be exclusively in float type. There could be instances of `strings` and `NaN` types, which need to be handled for certain tasks.

**Tasks:**

**1. Plot Stock Prices vs. Time for Two Companies in One Plot:**

**2. Plot Average Stock Price for Each Sector:**

Elaborating, suppose that there are three stocks: A, B, C. You are given the prices over a year. First compute the mean price in mu_A, mu_B, and mu_C over the year. 
Then, group the stocks by sector. For example, stocks A and B might be in the Tech sector and C might be in the Medical sector. Then, compute the average over all stocks in each sector. In this case, we compute mu_tech = (mu_A + mu_B) / 2

Which sector has historically been the best for investment?